Support binding array parameters
See: https://github.com/sripathikrishnan/jinjasql/pull/25
JinjaSQL made a bad assumption - an array or iterable cannot be a bind parameter.
Postgres and other modern databases allow arrays as bind parameters

JinjaSQL tried to be overly helpful, and raised an exception when it found array as a parameter.
There is no real need for us to raise MissingInClauseException.

Technically, this is a backwards incompatible change.
In practice though, the old behaviour would cause a MissingInclauseException, and now,
it will try to bind the parameter as an array. It should not cause a runtime problem.

Pull request #25 tried to add a new configuration parameter, so that this backward incompatible
change would not be needed. However, we feel that it is worth to release a new major version instead,
because the old behaviour was just wrong.